fix(lsp): handle missing client_id in error msg
`vim.lsp.buf.declaration()` results in an error due to a missing client id:

	bad argument #2 to 'format' (number expected, got nil)

This is due to the unsupported method check here:

	https://github.com/neovim/neovim/blob/968b4acd14568c9f9fed0dbace9527e3f7bbf493/runtime/lua/vim/lsp.lua#L1273-L1282

This check calls the resulting handler with the error message, but no
client_id because the error is not on a particular client, but rather no clients exist which support the method.

The solution is to omit the client id in the error message.