feat(ci/arm): add appimagetool + arch runtime to generate appimages
linuxdeploy doesn't seem to have a way to set/pass AppImage runtimes,
which we need to do since we're potentially building aarch64 on x86_64.

appimagetool has this option but is a lower level tool compared to
linuxdeploy, so for now we retain ld to prepare the AppDir and ait to
actually generate the image.