fix(tag): ex command "[count]tag" jumps to newer entries
Using the ex `:[count]tag` is supposed to jump to [count] newer
entry in the tagstack `:h tagstack`, this fixes the function `do_tag` to
actually follow this behavior. By refactoring the logic used in `:pop`
or CTRL-T to move to an item using the `tagstackidx`.

Previous to this changes jumping to newer entries in tagstack using
`:[count]tag` wasn't always possible because `do_tag` was searching for
the tag to then move to it, instead of moving forward in the tagstack.
Therefore unable to find tags that don't exist in a 'tagfile' or without
a 'tagfunc'. This is the case for tags defined using `settagstack` like
`vim.lsp.util.jump_to_location()` which is used for requests like:

- "textDocument/definition"
- "textDocument/declaration"
- "textDocument/typeDefinition"
- "textDocument/implementation"

Now there's a way to move to older items in the tagstack with `:pop` and
CTRL-T and to newer items in the tagstack with `:[count]tag`