vim-patch:8.2.2948: substitute() accepts a number but not a float expression
Problem:    Substitute() accepts a number but not a float expression.
Solution:   Also accept a float. (closes vim/vim#8331)
https://github.com/vim/vim/commit/7a2217bedd223df4c8bbebe731bf0b5fe8532533

Vim9script is N/A. No need to port the strict argument and
tv_get_string_buf_chk_strict, as it's only used for Vim9script.

This means similiar to numbers, floats are (mostly) convertable to
strings.