feat: add support for v:lua references in 'quickfixtextfunc'
The 'quickfixtextfunc' option works differently than other *func
variants such as 'omnifunc' and 'completefunc' because it parses its
value as a Callback struct, which currently does not support `v:lua`
function references. Callbacks support only function references or
partials, and partials are only supported when the name of the partial
function is present in the `partial_T` struct.

However, `v:lua` function references use a static `partial_T` struct.
This makes it impossible to use the callback machinery to call a `v:lua`
partial since we cannot mutate the single static `partial_T` struct
among multiple callback instances.

We get around this by introducing a new callback type
`kCallbackLuaFunc`. This is a blend of the `Funcref` and `Partial` types
in that it takes both a function name (the name of the Lua function) as
well as a `partial_T` struct (the static struct used for `v:lua`
references). This allows callbacks to specify a function name without
mutating the static struct.