refactor(lsp): move `response_to_list` to util and make public
Not really a feature, but I think it's useful to potentially have this
exposed in `util.lua`. For example in
https://github.com/scalameta/nvim-metals
we have a couple new features that I need to implement that are custom
endpoints in the server. These endpoints return `Location[]` so it'd be
really helpful to be able to just re-use this function, give it the
correct `entity` and call it a day. If not, I'll end up copying this
function to the plugin.