refactor(ex_eval): cherry-pick CSF flags changes from v8.2.3099
https://github.com/vim/vim/commit/4197828dc666f2d258594f7f9461534d23cc50e4

Cherry-pick the changes to existing flags values.
Required for v8.2.3478.

That patch mostly relates to Vim9script, but I'm careful not to mark it N/A in
case the flags have some use outside of Vim9 in the future.

Excludes CSF_FUNC_DEF (flag introduced in v8.2.1870 for Vim9's block scopes).