vim-patch:8.1.1925: more functions can be used as methods
Problem:    More functions can be used as methods.
Solution:   Make various functions usable as a method.
https://github.com/vim/vim/commit/4c313b13fb7bfa694ec6d2a13175e8650c007b2a

The rest of this patch was ported in
https://github.com/neovim/neovim/commit/3137c7d63574a86ddc44f11c839e8e58c2994bf9

N/A patches for version.c:

vim-patch:8.1.2012: more functions can be used as methods

Problem:    More functions can be used as methods.
Solution:   Make terminal functions usable as a method.  Fix term_getattr().
https://github.com/vim/vim/commit/7ee80f766130f22598c928415c92ec612a8f18cf

vim-patch:8.2.3576: some functions are not documented for use with a method

Problem:    Some functions are not documented for use with a method.
Solution:   Add examples.  Fix that sign_unplacelist() only takes one
            argument. (Sean Dewar, closes vim/vim#9081)
https://github.com/vim/vim/commit/a63780b9770699bf8fda93affeab2c728bc76912

Addressed in the port of v8.1.1995.