fix(ts): escape lang when loading parsers
When trying to load a language parser, escape the value of
the language.

With language injection, the language might be picked up from the
buffer. If this value is erroneous it can cause nvim_get_runtime_file
to hard error.

E.g. The markdown expression `~~~{` will extract '{' as a language and
then try to get the parser using `parser/{*` as the pattern.