fix(diagnostic): assert that diagnostics have line number and column
Line number and column are required and much of the diagnostic API
assumes that these are both present. When one of the two is missing,
cryptic errors pop up in other parts of the diagnostic subsystem.
Instead, assert that diagnostics are well formed when they are entered
into the cache, which provides a clearer error.

(cherry picked from commit 5399f862f2fc6c0ee435c2bfad01143c96af3db8)