fix: allow nvim -es to be used interactively with a TTY
When nvim -es is specified, the user enters ex mode and we don't enable
the terminal UI.  In the modern day, the user is typically using this
for scripting, but this also serves an important functionality as an
editor of last resort when the user finds themselves without a terminal.
ex mode is fully functional even on dumb terminals or when the terminfo
database is broken, and this can allow a user to recover their system.

However, there's no reason that a user shouldn't be able to use this
more interactively even if they do have a TTY.  This is not a common
thing to want to do, but it should be possible.  At the moment, though,
Neovim breaks this use case by explicitly checking for a missing
terminal before enabling input.  As a result, since we've enabled silent
mode, we have no terminal UI to read data from and we don't read from
standard input, so we just silently read nothing, and exit.

This is definitely not what the user wanted or what they asked for, and
it's bizarre to make this work only when the user is missing a TTY, so
let's just skip the TTY check altogether here so it works the same
regardless.