refactor: enable filetype detection before user startup scripts
The Nvim startup sequence is:

1. Source init.vim/init.lua
2. Source filetype.vim, ftplugin.vim, and indent.vim
3. Source syntax.vim

This PR deals with the ordering between 1 and 2. The problem with this
ordering is that any FileType autocommands created in 1 have a lower
priority than the ones created in 2. This is pertinent when people wish
to modify options globally that are often set by individual ftplugins (a
common example is 'formatoptions').

Concretely, if a user adds the following to their init.vim:

  au FileType * set formatoptions-=o

and then edits a file whose ftplugin sets 'formatoptions' with

  setlocal formatoptions+=croql

then the user's setting is overriden by the ftplugin. This is a common
source of confusion and frustration. The solution today is to use a
plugin file; that is, to put the FileType autocommand in
'plugin/autocmds.vim' or something similar.

This PR reverses the ordering of 1 and 2 so that the ftplugin
autocommands are created before the user's startup script is sourced.
This means any FileType autocommands in init.vim/init.lua will run
*after* ftplugin files, and users are free to override settings set by
the ftplugins.

The downside of this approach is that this removes the ability for users
to skip loading ftplugin files by adding

  filetype off

to their init file. Today, such a line will prevent the filetype.vim and
ftplugin.vim files being sourced at all. With this PR, filetype
detection will still be enabled, but users will still pay the "startup
cost".

I would argue that this is an extremely uncommon use case relative to
that of adding FileType autocommands to the init script, so this is an
acceptable trade-off.