feat: honor $XDG_RUNTIME_DIR when present
The XDG Base Directory Specification [1] suggests to store runtime files
below $XDG_RUNTIME_DIR when this environment variable is set. Currently,
neovim defaults to using a directory below $TMPDIR or /tmp. Make the
XDG_RUNTIME_DIR the preferred place for the temporary directory, but
fall back to the other locations when this environment variable is not
set.

[1]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
Ref: neovim/neovim#3517