fix(treesitter): correctly handle query quantifiers
Query patterns can contain quantifiers (e.g. (foo)+ @bar), so a single
capture can map to multiple nodes. The iter_matches API can not handle
this situation because the match table incorrectly maps capture indices
to a single node instead of to an array of nodes.

The match table should be updated to map capture indices to an array of
nodes. However, this is a massively breaking change, so must be done
with a proper deprecation period.

For now, `iter_matches` is deprecated and replaced with `iter_matches2`
which behaves correctly. This means that plugins must opt-in to correct
behavior. In 0.11, `iter_matches` will be hard deprecated, and in 0.12
it will be removed completely.

`add_predicate` and `add_directive` also must use an opt-in for correct
behavior to support backward compatibility. This is done with a new
"correct" option. This too will be removed in 0.12 and the correct
behavior will become the default.

Co-authored-by: Gregory Anders <greg@gpanders.com>