ci: create automated PRs as drafts
Now that the "lint-commits" check is required to merge, we need it to be
triggered before merging the automated PRs.  However, PRs
created/updated by workflows cannot trigger other workflows.

Therefore, create the PRs as drafts and remove them from draft status
when they're ready to merge.  Since that is a manual action, it will run
the lint-commit checker.

[skip ci]